// types.ts
export interface PostProfile {
  full_name: string | null;
  avatar_url: string | null;
}
// ...
export interface Group {
    id: string;
    created_at: string;
    name: string;
    description: string | null;
    cover_image_url: string | null;
    is_private: boolean;
    user_id: string; // owner
    profiles: PostProfile | null; // owner profile
    group_members: [{ count: number }];
}